Split raw GL/Skia FBO interop out of SkiaGlBackend into engine-agnostic Core.OGL
First step of #3: SkiaGlBackend's framebuffer/renderbuffer plumbing was pure
OpenGL with no real MonoGame dependency, but lived inline in a MonoGame-coupled
project. Moving it to SkiaMonoGameRendering.Core.OGL (SkiaSharp-only, no engine
reference) means the next GL-based backend (raylib) can reuse it instead of
duplicating it. SkiaGlBackend keeps all the MonoGame/SDL-specific glue (window
handle discovery, shared context creation, texture handle capture) and now
just calls into Core.OGL for surface creation.

No behavior change: verified via the existing Tests.Core suite and by running
Sample.MonoGame.DesktopGL to confirm Skia content still renders correctly.
