+++ SkiaMonoGameRendering.Core.OGL/SkiaMonoGameRendering.Core.OGL.csproj
@@ -8,12 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="MonoGame.Framework.DesktopGL" Version="3.8.4.1" />
     <PackageReference Include="SkiaSharp" Version="3.119.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SkiaMonoGameRendering.Core.OGL\SkiaMonoGameRendering.Core.OGL.csproj" />
   </ItemGroup>
 
 </Project>
